avance en cambios del master

+++ app_1_addition/models.py
@@ -129,7 +129,7 @@ class Player(BasePlayer):
         print("[[ APP_1_ADDITION]] - PLAYER - FINAL_COUNT.............[[[ PLAYER_ID ==> ", self.id_in_group, " <== ]]]")
 
         self.acc_was_correct = sum(filter(None, [p.was_correct for p in self.in_all_rounds()]))
-        self.acc_payoff = sum([i * Constants.piece_rate for i in [p.was_correct for p in self.in_all_rounds()] if i != None])  # this creates a list multiplying every correct '1' times the piece rate and then adds it all
+        self.acc_payoff = int(sum([i * Constants.piece_rate for i in [p.was_correct for p in self.in_all_rounds()] if i != None]))  # this creates a list multiplying every correct '1' times the piece rate and then adds it all
 
         if self.participant.vars['treatment'] == 1:
             self.final_payoff = self.acc_payoff
